requirements.txt
anthropic
chainlit
python-dotenv
